Move tileApi registration up into graph-wrapper-component.
