+++ src/plugins/graph/components/graph-component.test.tsx
@@ -3,22 +3,24 @@ import userEvent from "@testing-library/user-event";
 import React from "react";
 import { DataBroker } from "../../../models/data/data-broker";
 import { DataSet } from "../../../models/data/data-set";
+import { GraphLayout } from "../models/graph-layout";
 import { GraphComponent } from "./graph-component";
 
 describe.skip("Graph", () => {
   let broker: DataBroker;
+  const layout = new GraphLayout();
   beforeEach(() => {
     broker = new DataBroker();
   });
 
   it("renders with no broker", () => {
-    render(<GraphComponent />);
+    render(<GraphComponent layout={layout} />);
     // expect(screen.getByTestId("graph")).toBeInTheDocument()
     expect(true).toBe(true);
   });
 
   it("renders with empty broker", () => {
-    render(<GraphComponent/>);
+    render(<GraphComponent layout={layout} />);
     expect(screen.getByTestId("graph")).toBeInTheDocument();
   });
 
@@ -28,7 +30,7 @@ describe.skip("Graph", () => {
     data.addAttributeWithID({ name: "yVariable" });
     data.addCasesWithIDs([{ xVariable: 1, yVariable: 2, __id__: "c1" }, { xVariable: 3, yVariable: 4, __id__: "c2" }]);
     broker.addDataSet(data);
-    render(<GraphComponent/>);
+    render(<GraphComponent layout={layout} />);
     expect(screen.getByTestId("graph")).toBeInTheDocument();
     // rerender(<GraphComponent broker={broker} />)
     // expect(screen.getByText('xVariable')).toBeInTheDocument()
@@ -42,12 +44,12 @@ describe.skip("Graph", () => {
     data.addAttributeWithID({ name: "yVariable" });
     data.addCasesWithIDs([{ xVariable: 1, yVariable: 2, __id__: "c1" }, { xVariable: 3, yVariable: 4, __id__: "c2" }]);
     broker.addDataSet(data);
-    const { rerender } = render(<GraphComponent/>);
+    const { rerender } = render(<GraphComponent layout={layout} />);
     expect(screen.getByTestId("graph")).toBeInTheDocument();
     // expect(screen.getByText('xVariable')).toBeInTheDocument()
     // expect(screen.getByText('yVariable')).toBeInTheDocument()
     const plotTypeButton = screen.getByText("Dot Plot");
     await user.click(plotTypeButton);
-    rerender(<GraphComponent/>);
+    rerender(<GraphComponent layout={layout} />);
   });
 });
